xss/requirements.txt
flask
apscheduler
selenium
mysql-connector-python